requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.1.0
Django==4.2
django-admin-thumbnails==0.2.7
idna==3.4
Pillow==9.5.0
requests==2.29.0
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==1.26.15
xlwt==1.3.0
